Handle case where the view is not present in the current view
